api/pyproject.toml
[tool.poetry]
name = "osint-api"
version = "0.1.0"
description = "OSINT Platform API"
authors = ["OSINT Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.25"
psycopg = {extras = ["binary"], version = "^3.1.0"}
alembic = "^1.13.0"
redis = "^5.0.0"
pydantic-settings = "^2.1.0"
structlog = "^24.1.0"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
ruff = "^0.4.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
